# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enspatium API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "enspatium_session"
      }
    },
...
  },
  "paths": {
...
    "/namespaces": {
      "post": {
        "operationId": "createOrganizationNamespace",
        "tags": [
          "namespaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "slug": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "ownerUserId",
                    "name",
                    "slug",
                    "kind",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "ownerUserId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "kind": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "personal"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "organization"
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listNamespaces",
        "tags": [
          "namespaces"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "ownerUserId",
                      "name",
                      "slug",
                      "kind",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "ownerUserId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "kind": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "personal"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "organization"
                            ]
                          }
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/namespaces/{slug}": {
      "get": {
        "operationId": "getNamespace",
        "tags": [
          "namespaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "ownerUserId",
                    "name",
                    "slug",
                    "kind",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "ownerUserId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "kind": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "personal"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "organization"
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/namespaces/{slug}/members": {
      "post": {
        "operationId": "addNamespaceMember",
        "tags": [
          "namespaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "minLength": 1,
                    "maxLength": 320,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "userId",
                    "email",
                    "displayName",
                    "role",
                    "joinedAt"
                  ],
                  "properties": {
                    "userId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "member"
                          ]
                        }
                      ]
                    },
                    "joinedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listNamespaceMembers",
        "tags": [
          "namespaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "userId",
                      "email",
                      "displayName",
                      "role",
                      "joinedAt"
                    ],
                    "properties": {
                      "userId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      },
                      "role": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "owner"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "member"
                            ]
                          }
                        ]
                      },
                      "joinedAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/namespaces/{slug}/members/{userId}": {
      "delete": {
        "operationId": "removeNamespaceMember",
        "tags": [
          "namespaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ]
}